scanner adds isbn to search
